[2023.02.08 노영준] 블루팀 구조물 추가

+++ app/src/main/java/com/choonham/lck_manager/map_object/Champion.java
@@ -2,7 +2,7 @@ package com.choonham.lck_manager.map_object;
 
 public class Champion extends MapObject{
 
-    public Champion(String objectName, int x, int y, int visionDistance, int teamSide) {
+    public Champion(String objectName, float x, float y, int visionDistance, int teamSide) {
         super(objectName, x, y, visionDistance, teamSide);
     }
 }
