+++ app/src/main/java/com/choonham/lck_manager/map_object/Inhibitor.java
@@ -2,7 +2,7 @@ package com.choonham.lck_manager.map_object;
 
 public class Inhibitor extends MapObject{
 
-    public Inhibitor(String objectName, int x, int y, int visionDistance, int teamSide) {
+    public Inhibitor(String objectName, float x, float y, int visionDistance, int teamSide) {
         super(objectName, x, y, visionDistance, teamSide);
     }
 }
