+++ app/src/main/java/com/choonham/lck_manager/map_object/Nexus.java
@@ -2,7 +2,7 @@ package com.choonham.lck_manager.map_object;
 
 public class Nexus extends MapObject{
 
-    public Nexus(String objectName, int x, int y, int visionDistance, int teamSide) {
+    public Nexus(String objectName, float x, float y, int visionDistance, int teamSide) {
         super(objectName, x, y, visionDistance, teamSide);
     }
 }
