+++ app/src/main/java/com/choonham/lck_manager/map_object/Turret.java
@@ -2,7 +2,7 @@ package com.choonham.lck_manager.map_object;
 
 public class Turret  extends MapObject{
 
-    public Turret(String objectName, int x, int y, int visionDistance, int teamSide) {
+    public Turret(String objectName, float x, float y, int visionDistance, int teamSide) {
         super(objectName, x, y, visionDistance, teamSide);
     }
